Move PhoneState.kt to devicemanagement/wear
