+++ devicemanagement/wear/src/androidMain/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.boswelja.smartwatchextensions.devicemanagement" />
