+++ wearos/src/main/kotlin/com/boswelja/smartwatchextensions/main/ui/MainViewModel.kt
@@ -3,8 +3,8 @@ package com.boswelja.smartwatchextensions.main.ui
 import androidx.datastore.core.DataStore
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.boswelja.smartwatchextensions.PhoneState
 import com.boswelja.smartwatchextensions.capability.CapabilityUpdater
+import com.boswelja.smartwatchextensions.devicemanagement.PhoneState
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
